pyproject.toml
[project]
name = "nonebot_plugin_zikequote3"
version = "0.3.2"
description = "LLM 介入，功能丰富的群聊语录自动收集与管理插件，支持自动收集、后台管理、生成排行榜、展示等功能"
authors = [
  { name="LifeCheckpoint" },
]
dependencies = [
  "nonebot2>=2.0.0",
  "nonebot-adapter-onebot>=2.0.0",
  "nonebot-plugin-localstore>=0.7.0",
  "pydantic>=2.0.0",
  "openai>=1.76.0",
  "requests>=2.26.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=1.0.0",
  "nonebug>=0.4.3",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text="MIT" }

[project.urls]
Homepage = "https://github.com/LifeCheckpoint/ZikeQuote3"

[tool.nonebot]
plugins = ["ZikeQuote3"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.setuptools]
packages = ["nonebot_plugin_zikequote3"]